requirements.txt
fastapi
uvicorn
requests
pandas
openai
pillow
markdown2
beautifulsoup4
faker
easyocr